+++ jaxley/modules/network.py
@@ -240,6 +240,9 @@ class Network(Module):
         self._indices_jax_spsolve = indices
         self._indptr_jax_spsolve = indptr
 
+        # To enable updating `self._comp_edges` during `View`.
+        self._comp_edges_in_view = self._comp_edges.index.to_numpy()
+
     def _step_synapse(
         self,
         states: Dict,